Cover inherits in remaining language examples
