+++ examples/example-objective-c/README.md
@@ -10,6 +10,8 @@ Open `examples/example-objective-c` in CodeGraphy and look for:
 - `Sources/Controllers/DashboardController.m -> Sources/Data/SessionStore.h#import`
 - `Sources/Controllers/DashboardController.m -> Sources/Feature/UserCardView.h#import`
 - `Sources/Feature/UserCardView.m -> Sources/Models/UserProfile.h#import`
+- `Sources/Feature/UserCardView.h -> Sources/Feature/AppView.h#inherit`
+- `Sources/Feature/UserCardView.h -> Sources/Feature/ProfileRenderable.h#inherit`
 
 ## Symbol Node Demo
 
@@ -17,9 +19,10 @@ Suggested symbol check:
 
 1. Open `Sources/AppDelegate.m`.
 2. In Graph Scope, enable **Symbol**.
-3. Search for `main`, `AppDelegate`, `DashboardController`, `UserCardView`, and `renderTitle`.
+3. Search for `main`, `AppDelegate`, `DashboardController`, `AppView`, `ProfileRenderable`, `UserCardView`, and `renderTitle`.
 
 Expected behavior:
 
 - Interface and implementation declarations appear as class symbols.
 - Objective-C method declarations/definitions appear as method symbols.
+- `UserCardView` inherits from `AppView` and conforms to `ProfileRenderable`, giving Objective-C inheritance and protocol conformance distinct file targets.
