+++ examples/example-objective-c/Sources/Feature/AppView.h
@@ -1,0 +1,4 @@
+#import <Foundation/Foundation.h>
+
+@interface AppView : NSObject
+@end
