+++ examples/example-objective-c/Sources/Feature/ProfileRenderable.h
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+
+@protocol ProfileRenderable
+- (NSString *)renderTitle;
+@end
