+++ examples/example-python/src/services/base.py
@@ -1,0 +1,8 @@
+"""Base API service records."""
+
+
+class BaseApiUser:
+    """Shared API user shape."""
+
+    def __init__(self, name):
+        self.name = name
