+++ examples/example-svelte/src/inheritance.ts
@@ -1,0 +1,3 @@
+export interface DisplayEntity {
+  name: string;
+}
