+++ examples/example-typescript/src/baseRunner.ts
@@ -1,0 +1,5 @@
+export class BaseRunner {
+  start(): string {
+    return 'ready';
+  }
+}
