+++ examples/example-typescript/src/runnableThing.ts
@@ -1,0 +1,3 @@
+export interface RunnableThing {
+  run(): string;
+}
