+++ examples/example-typescript/src/runner.ts
@@ -1,0 +1,8 @@
+import { BaseRunner } from './baseRunner';
+import type { RunnableThing } from './runnableThing';
+
+export class AppRunner extends BaseRunner implements RunnableThing {
+  run(): string {
+    return this.start();
+  }
+}
