+++ examples/example-vue/src/types.ts
@@ -1,9 +1,10 @@
+import type { DisplayEntity } from './inheritance';
+
 export interface CounterPanelProps {
   title: string;
   count: number;
 }
 
-export interface UserProfile {
-  name: string;
+export interface UserProfile extends DisplayEntity {
   role: string;
 }
