+++ packages/extension/tests/acceptance/specs/godot-example.md
@@ -28,3 +28,10 @@ And scripts/orphan.gd is an orphan node
 And README.md is an orphan node
 And .gitignore is an orphan node
 And .vscode/settings.json is an orphan node
+
+When I show only the File node type
+And I show no edge types
+Then the top right of the graph says "0 connections"
+When I toggle the Inherits edge on
+Then the top right of the graph says "1 connection"
+And scripts/enemy.gd points to scripts/base/entity.gd
